+++ src/main/java/com/bookbla/americano/domain/library/controller/MemberLibraryController.java
@@ -2,9 +2,9 @@ package com.bookbla.americano.domain.library.controller;
 
 import com.bookbla.americano.base.resolver.LoginUser;
 import com.bookbla.americano.base.resolver.User;
-import com.bookbla.americano.domain.library.service.MemberLibraryService;
 import com.bookbla.americano.domain.library.controller.dto.MemberLibraryProfileReadResponse;
 import com.bookbla.americano.domain.library.controller.dto.MemberTargetLibraryProfileReadResponse;
+import com.bookbla.americano.domain.library.service.MemberLibraryService;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
