+++ src/main/java/com/bookbla/americano/domain/member/service/StyleService.java
@@ -1,8 +1,26 @@
 package com.bookbla.americano.domain.member.service;
 
 import com.bookbla.americano.domain.member.controller.dto.response.StylesResponse;
+import com.bookbla.americano.domain.member.enums.ContactType;
+import com.bookbla.americano.domain.member.enums.DateCostType;
+import com.bookbla.americano.domain.member.enums.DateStyleType;
+import com.bookbla.americano.domain.member.enums.DrinkType;
+import com.bookbla.americano.domain.member.enums.JustFriendType;
+import com.bookbla.americano.domain.member.enums.SmokeType;
+import org.springframework.stereotype.Service;
 
-public interface StyleService {
+@Service
+public class StyleService {
 
-    StylesResponse readStyles();
+    public StylesResponse readStyles() {
+        return StylesResponse.of(
+                SmokeType.getValues(),
+                DrinkType.getValues(),
+                ContactType.getValues(),
+                DateStyleType.getValues(),
+                DateCostType.getValues(),
+                JustFriendType.getValues()
+        );
+    }
+
 }
